Implement ApkMirror updates Add setting for number of columns

+++ app/src/main/java/com/apkupdater/data/apkmirror/AppExistsResponseApk.kt
@@ -3,7 +3,7 @@ package com.apkupdater.data.apkmirror
 import com.google.gson.annotations.SerializedName
 
 data class AppExistsResponseApk(
-	@SerializedName("version_code") val versionCode: Int = 0,
+	@SerializedName("version_code") val versionCode: Long = 0,
 	val link: String = "",
 	@SerializedName("publish_date") val publishDate: String? = null,
 	val arches: List<String> = emptyList(),
